LayoutTests/fast/text/font-synthesis-parsing.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<div id="test1" style="font-synthesis: weight;"></div>
<div id="test2" style="font-synthesis: small-caps;"></div>
<div id="test3" style="font-synthesis: none;"></div>
<div id="test4" style="font-synthesis: garbage;"></div>
<div id="test5" style="font-synthesis: initial;"></div>
<div id="test6" style="font-synthesis: style small-caps;"></div>
<div id="test7" style="font-synthesis: small-caps style;"></div>
<script>
shouldBeTrue("CSS.supports('font-synthesis', 'small-caps')");
shouldBeEqualToString("window.getComputedStyle(document.getElementById('test1')).fontSynthesis", "weight");
shouldBeEqualToString("window.getComputedStyle(document.getElementById('test2')).fontSynthesis", "small-caps");
shouldBeEqualToString("window.getComputedStyle(document.getElementById('test3')).fontSynthesis", "none");
shouldBeEqualToString("window.getComputedStyle(document.getElementById('test4')).fontSynthesis", "style weight small-caps");
shouldBeEqualToString("window.getComputedStyle(document.getElementById('test5')).fontSynthesis", "style weight small-caps");
shouldBeEqualToString("window.getComputedStyle(document.getElementById('test6')).fontSynthesis", "style small-caps");
shouldBeEqualToString("window.getComputedStyle(document.getElementById('test7')).fontSynthesis", "style small-caps");
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
