LayoutTests/fast/text/font-synthesis-parsing-expected.txt

PASS CSS.supports('font-synthesis', 'small-caps') is true
PASS window.getComputedStyle(document.getElementById('test1')).fontSynthesis is "weight"
PASS window.getComputedStyle(document.getElementById('test2')).fontSynthesis is "small-caps"
PASS window.getComputedStyle(document.getElementById('test3')).fontSynthesis is "none"
PASS window.getComputedStyle(document.getElementById('test4')).fontSynthesis is "style weight small-caps"
PASS window.getComputedStyle(document.getElementById('test5')).fontSynthesis is "style weight small-caps"
PASS window.getComputedStyle(document.getElementById('test6')).fontSynthesis is "style small-caps"
PASS window.getComputedStyle(document.getElementById('test7')).fontSynthesis is "style small-caps"
PASS successfullyParsed is true

TEST COMPLETE
